// src/models/IoK8sApimachineryPkgApisMetaV1APIGroupList/types.ts
import { IoK8sApimachineryPkgApisMetaV1APIGroupModelOptions as IoK8sApimachineryPkgApisMetaV1APIGroup } from '../IoK8sApimachineryPkgApisMetaV1APIGroup';

export interface IoK8sApimachineryPkgApisMetaV1APIGroupListModelOptions {
  apiVersion?: string;
  groups: Array<IoK8sApimachineryPkgApisMetaV1APIGroup>;
  kind?: string;
}